+++ .gitignore
@@ -8,3 +8,4 @@ IndividualData
 PCEPI.csv
 psid.xlsx
 WealthData
+.ipynb_checkpoints/
