--- __init__.py
@@ -1,6 +1,0 @@
-import pandas as pd
-import numpy as np
-import psidraw
-import psiddata
-import psidvars
-from functions import *
